Integrate 4-byte alignment optimization into main int8 CUDA kernel
- Updated main int8 kernel with conditional aligned uchar4 loads
  - Runtime alignment detection for optimal memory access
  - Falls back gracefully for non-aligned data
  - Provides 1.6x speedup (59.5% improvement) for aligned patterns

- Enhanced synthetic pattern generation with automatic alignment
  - Modified block positioning to round x-coordinates to multiples of 4
  - Maintains pattern integrity while ensuring optimal performance
  - Applied across all positioning strategies (pattern-based, LED-based, fallback)

- Added comprehensive test suite for alignment experiments
  - Performance comparison between aligned and non-aligned patterns
  - Validation of mathematical equivalence between kernel variants
  - Focused alignment experiment demonstrating performance benefits

- Generated new aligned synthetic patterns (1000 LEDs, 100% aligned)
  - All x-positions verified as multiples of 4
  - 284MB pattern file with sparse matrix and mixed tensor formats

Performance improvements:
- Random positions (27% aligned): 3.623ms → Aligned positions (100%): 2.272ms
- Experimental kernel comparison shows up to 3.38x speedup with full alignment
- Backward compatible with existing non-aligned diffusion patterns

+++ tests/test_single_block_sparse_tensor.py
@@ -222,10 +222,10 @@ def test_save_load():
 
     # Set some blocks with known values
     test_data = [
-        (0, 0, 5, 10, cp.ones((8, 8)) * 1.5),
-        (1, 1, 15, 20, cp.ones((8, 8)) * 2.5),
-        (3, 0, 25, 5, cp.ones((8, 8)) * 3.5),
-        (7, 1, 10, 35, cp.ones((8, 8)) * 4.5),
+        (0, 0, 5, 10, cp.ones((8, 8), dtype=cp.float32) * 1.5),
+        (1, 1, 15, 20, cp.ones((8, 8), dtype=cp.float32) * 2.5),
+        (3, 0, 25, 5, cp.ones((8, 8), dtype=cp.float32) * 3.5),
+        (7, 1, 10, 35, cp.ones((8, 8), dtype=cp.float32) * 4.5),
     ]
 
     for batch_idx, channel_idx, row, col, values in test_data:
